fix: 변수명 camelcase로 수정

+++ src/main/java/com/goggle/voco/dto/AudioInputResponseDto.java
@@ -10,5 +10,5 @@ import lombok.Setter;
 @NoArgsConstructor
 @AllArgsConstructor
 public class AudioInputResponseDto {
-    private int textId;
+    private Long textId;
 }
